Refactor: Move TradingConfig and improve validation
