+++ live_trading/__init__.py
@@ -6,13 +6,14 @@ including order management, data streaming, and exchange integration.
 """
 
 from .trading_manager import TradingManager
-from .order_manager import OrderManager
+from .order_manager import OrderManager, TradingConfig
 from .data_receiver import DataReceiver
 from .trade_executor import TradeExecutor
 
 __all__ = [
     "TradingManager",
     "OrderManager",
+    "TradingConfig",
     "DataReceiver",
     "TradeExecutor"
 ]
